+ make plink.mk more robust
  - ensure that the Makefile "successfully" works when the input vcf
    contains no variants
     - check for an empty input vcf
     - handle the empty input vcf case accordingly (ensure a "empty-vcf"
       file is created
     - make the reformat functionality a function instead of a PHONY
       target
     - use the bash shell
   - ensure that the output directory gets created no matter what

+++ yaps2/resources/mie/plink.mk
@@ -1,4 +1,7 @@
-.PHONY: clean reformat check-env
+.PHONY: clean reformat check-env create-project-dir
+
+# set the shell to use bash
+SHELL := /bin/bash
 
 # needed inputs
 INPUT_VCF :=                # aka "trios.vcf.gz"
@@ -30,34 +33,62 @@ unfiltered-mendel := $(unfiltered).mendel
 unfiltered-fmendel := $(unfiltered).fmendel
 unfiltered-imendel := $(unfiltered).imendel
 unfiltered-lmendel := $(unfiltered).lmendel
+empty_vcf := $(PRJ_DIR)/empty-vcf
 
-all: check-env $(unfiltered-mie-var-txt) $(unfiltered-info-variants-var)
+INPUT_VCF_VARIANT_COUNT := $(shell $(BCFTOOLS) view $(INPUT_VCF) | grep -v '^\#' | head -n 1 | wc -l )
 
-$(unfiltered-mie-var-txt): $(unfiltered-info-variants-var) $(unfiltered-fmendel) reformat
+ifeq ($(shell test $(INPUT_VCF_VARIANT_COUNT) -ge 1; echo $$?),0)
+EMPTY_INPUT_VCF_FLAG :=
+else
+EMPTY_INPUT_VCF_FLAG := "empty"
+endif
+
+all: check-env create-project-dir $(unfiltered-mie-var-txt)
+
+$(unfiltered-mie-var-txt): $(unfiltered-info-variants-var) $(unfiltered-fmendel)
+ifndef EMPTY_INPUT_VCF_FLAG
 	$(BASH) $(MAKE_TABLE) $(unfiltered-info-variants-var) $(unfiltered-fmendel) > $(unfiltered-mie-var-txt)
+else
+	echo "$(INPUT_VCF) has no variants!"
+	touch $(empty_vcf)
+endif
 
 $(unfiltered-info-variants-var): $(TRIO_FAM) $(INPUT_VCF)
 	$(BASH) $(SUMMARIZE_INFORMATIVE_VAR_ONLY) $(TRIO_FAM) ${INPUT_VCF} > $(unfiltered-info-variants-var)
 
 $(unfiltered-mendel) $(unfiltered-fmendel) $(unfiltered-imendel) $(unfiltered-lmendel): $(unfiltered-bed) $(unfiltered-bim) $(unfiltered-fam)
+ifndef EMPTY_INPUT_VCF_FLAG
 	# excluding anything that isn't on an autosome as I suspect XY etc are not well handled
 	$(PLINK) --bfile $(unfiltered) --mendel --allow-extra-chr --out $(unfiltered) --chr 1-22
+	$(eval PLINKOUT := $(unfiltered-mendel) $(unfiltered-fmendel) $(unfiltered-imendel) $(unfiltered-lmendel))
+	$(call reformat,$(PLINKOUT))
+else
+	touch $(unfiltered-fmendel)
+endif
 
 $(unfiltered-bed) $(unfiltered-bim) $(unfiltered-fam): $(INPUT_VCF) $(TRIO_FAM)
+ifndef EMPTY_INPUT_VCF_FLAG
 	# allow-extra-chr is there to allow unplaced contigs
 	# double id is a reasonable default, but we will replace the ouput .fam file with our own later below
 	$(PLINK) --vcf $(INPUT_VCF) --double-id --make-bed --out $(unfiltered) --allow-extra-chr
 	$(RM) $(unfiltered-nosex)
 	$(CP) $(TRIO_FAM) $(unfiltered-fam)
+else
+	touch $(unfiltered-bed) $(unfiltered-bim) $(unfiltered-fam)
+	$(CP) $(TRIO_FAM) $(unfiltered-fam)
+endif
 
-PLINKOUT = $(unfiltered-mendel) $(unfiltered-fmendel) $(unfiltered-imendel) $(unfiltered-lmendel)
-reformat:
+define reformat
 	# This takes each plink output file and translates it to use tabs.
 	# Commands taken from https://www.cog-genomics.org/plink2/other#tabspace
-	$(foreach file,$(PLINKOUT),cat $(file) \
+	$(foreach file,$1,cat $(file) \
 		| sed 's/^[[:space:]]*//g' \
 		| sed 's/[[:space:]]*$$//g' \
 		| tr -s ' ' '\t' > $(file).tmp && mv $(file).tmp $(file);)
+endef
+
+create-project-dir:
+	mkdir -p $(PRJ_DIR)
 
 clean: check-env
 	$(RM) $(PRJ_DIR)
